Moved MainPage to handler files, doc update
